+++ lib/src/utils/platform_helper.dart
@@ -1,3 +1,0 @@
-import 'dart:io';
-
-bool isMobile() => Platform.isAndroid || Platform.isIOS;