+++ lib/src/widgets/link.dart
@@ -2,7 +2,26 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../../models/documents/nodes/node.dart';
+import '../models/documents/attribute.dart';
+import '../models/documents/nodes/node.dart';
+import '../translations/toolbar.i18n.dart';
+
+const linkPrefixes = [
+  'mailto:', // email
+  'tel:', // telephone
+  'sms:', // SMS
+  'callto:',
+  'wtai:',
+  'market:',
+  'geopoint:',
+  'ymsgr:',
+  'msnim:',
+  'gtalk:', // Google Talk
+  'skype:',
+  'sip:', // Lync
+  'whatsapp:',
+  'http'
+];
 
 /// List of possible actions returned from [LinkActionPickerDelegate].
 enum LinkMenuAction {
@@ -40,6 +59,33 @@ Future<LinkMenuAction> defaultLinkActionPickerDelegate(
           'be invoked for $defaultTargetPlatform');
       return LinkMenuAction.none;
   }
+}
+
+TextRange getLinkRange(Node node) {
+  var start = node.documentOffset;
+  var length = node.length;
+  var prev = node.previous;
+  final linkAttr = node.style.attributes[Attribute.link.key]!;
+  while (prev != null) {
+    if (prev.style.attributes[Attribute.link.key] == linkAttr) {
+      start = prev.documentOffset;
+      length += prev.length;
+      prev = prev.previous;
+    } else {
+      break;
+    }
+  }
+
+  var next = node.next;
+  while (next != null) {
+    if (next.style.attributes[Attribute.link.key] == linkAttr) {
+      length += next.length;
+      next = next.next;
+    } else {
+      break;
+    }
+  }
+  return TextRange(start: start, end: start + length);
 }
 
 Future<LinkMenuAction> _showCupertinoLinkMenu(
@@ -126,18 +172,18 @@ Future<LinkMenuAction> _showMaterialMenu(
             child: Text(link),
           ),
           _MaterialAction(
-            title: 'Open',
+            title: 'Open'.i18n,
             icon: Icons.language_sharp,
             onPressed: () => Navigator.of(context).pop(LinkMenuAction.launch),
           ),
           _MaterialAction(
-            title: 'Copy',
+            title: 'Copy'.i18n,
             icon: Icons.copy_sharp,
             onPressed: () => Navigator.of(context).pop(LinkMenuAction.copy),
           ),
           if (showDelete)
             _MaterialAction(
-              title: 'Remove',
+              title: 'Remove'.i18n,
               icon: Icons.link_off_sharp,
               onPressed: () => Navigator.of(context).pop(LinkMenuAction.remove),
             ),