+++ lib/src/widgets/raw_editor/raw_editor_state_text_input_client_mixin.dart
@@ -2,10 +2,11 @@ import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 
 import '../../models/documents/document.dart';
-import '../../utils/diff_delta.dart';
+import '../../utils/delta.dart';
 import '../editor.dart';
 
 mixin RawEditorStateTextInputClientMixin on EditorState
@@ -60,8 +61,8 @@ mixin RawEditorStateTextInputClientMixin on EditorState
         ),
       );
 
+      _updateSizeAndTransform();
       _textInputConnection!.setEditingState(_lastKnownRemoteTextEditingValue!);
-      // _sentRemoteValues.add(_lastKnownRemoteTextEditingValue!);
     }
 
     _textInputConnection!.show();
@@ -110,6 +111,7 @@ mixin RawEditorStateTextInputClientMixin on EditorState
     );
   }
 
+  // Start TextInputClient implementation
   @override
   TextEditingValue? get currentTextEditingValue =>
       _lastKnownRemoteTextEditingValue;
@@ -295,5 +297,20 @@ mixin RawEditorStateTextInputClientMixin on EditorState
     _textInputConnection!.connectionClosedReceived();
     _textInputConnection = null;
     _lastKnownRemoteTextEditingValue = null;
+  }
+
+  void _updateSizeAndTransform() {
+    if (hasConnection) {
+      // Asking for renderEditor.size here can cause errors if layout hasn't
+      // occurred yet. So we schedule a post frame callback instead.
+      SchedulerBinding.instance!.addPostFrameCallback((_) {
+        if (!mounted) {
+          return;
+        }
+        final size = renderEditor.size;
+        final transform = renderEditor.getTransformTo(null);
+        _textInputConnection?.setEditableSizeAndTransform(size, transform);
+      });
+    }
   }
 }