+++ lib/src/widgets/text_line.dart
@@ -1,23 +1,27 @@
 import 'dart:collection';
 import 'dart:math' as math;
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter_quill/flutter_quill.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:tuple/tuple.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../../flutter_quill.dart';
+import '../models/documents/attribute.dart';
 import '../models/documents/nodes/container.dart' as container_node;
+import '../models/documents/nodes/leaf.dart';
 import '../models/documents/nodes/leaf.dart' as leaf;
 import '../models/documents/nodes/line.dart';
 import '../models/documents/nodes/node.dart';
 import '../models/documents/style.dart';
 import '../utils/color.dart';
+import '../utils/platform.dart';
 import 'box.dart';
+import 'controller.dart';
 import 'cursor.dart';
+import 'default_styles.dart';
 import 'delegate.dart';
 import 'keyboard_listener.dart';
 import 'proxy.dart';
@@ -73,23 +77,17 @@ class _TextLineState extends State<TextLine> {
     }
   }
 
-  bool get isDesktop => {
-        TargetPlatform.macOS,
-        TargetPlatform.linux,
-        TargetPlatform.windows
-      }.contains(defaultTargetPlatform);
-
   bool get canLaunchLinks {
     // In readOnly mode users can launch links
     // by simply tapping (clicking) on them
     if (widget.readOnly) return true;
 
-    if (widget.hasFocus && !isDesktop) return false;
+    if (widget.hasFocus && !isDesktop()) return false;
     // In editing mode it depends on the platform:
 
     // Desktop platforms (macos, linux, windows):
     // only allow Meta(Control)+Click combinations
-    if (isDesktop) {
+    if (isDesktop()) {
       return _metaOrControlPressed || !widget.hasFocus;
     }
     // Mobile platforms (ios, android): always allow but we install a
@@ -400,7 +398,7 @@ class _TextLineState extends State<TextLine> {
       return _linkRecognizers[segment]!;
     }
 
-    // if (/*isDesktop ||*/ widget.hasFocus) {
+    // if (/*isDesktop() ||*/ widget.hasFocus) {
     _linkRecognizers[segment] = TapGestureRecognizer()
       ..onTap = () => _tapNodeLink(segment);
     // } // else {
@@ -462,42 +460,13 @@ class _TextLineState extends State<TextLine> {
         Clipboard.setData(ClipboardData(text: link));
         break;
       case LinkMenuAction.remove:
-        final range = _getLinkRange(node);
+        final range = getLinkRange(node);
         widget.controller
             .formatText(range.start, range.end - range.start, Attribute.link);
         break;
       case LinkMenuAction.none:
         break;
     }
-  }
-*/
-/*
-
-  TextRange _getLinkRange(Node node) {
-    var start = node.documentOffset;
-    var length = node.length;
-    var prev = node.previous;
-    final linkAttr = node.style.attributes[Attribute.link.key]!;
-    while (prev != null) {
-      if (prev.style.attributes[Attribute.link.key] == linkAttr) {
-        start = prev.documentOffset;
-        length += prev.length;
-        prev = prev.previous;
-      } else {
-        break;
-      }
-    }
-
-    var next = node.next;
-    while (next != null) {
-      if (next.style.attributes[Attribute.link.key] == linkAttr) {
-        length += next.length;
-        next = next.next;
-      } else {
-        break;
-      }
-    }
-    return TextRange(start: start, end: start + length);
   }
 */
 
@@ -895,19 +864,10 @@ class RenderEditableTextLine extends RenderEditableBox {
   /// of the cursor for iOS is approximate and obtained through an eyeball
   /// comparison.
   void _computeCaretPrototype() {
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        _caretPrototype = Rect.fromLTWH(0, 0, cursorWidth, cursorHeight + 2);
-        break;
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.linux:
-      case TargetPlatform.windows:
-        _caretPrototype = Rect.fromLTWH(0, 2, cursorWidth, cursorHeight - 4.0);
-        break;
-      default:
-        throw 'Invalid platform';
+    if (isAppleOS()) {
+      _caretPrototype = Rect.fromLTWH(0, 0, cursorWidth, cursorHeight + 2);
+    } else {
+      _caretPrototype = Rect.fromLTWH(0, 2, cursorWidth, cursorHeight - 4.0);
     }
   }