+++ lib/src/widgets/toolbar/color_button.dart
@@ -39,7 +39,7 @@ class _ColorButtonState extends State<ColorButton> {
   late bool _isToggledColor;
   late bool _isToggledBackground;
   late bool _isWhite;
-  late bool _isWhitebackground;
+  late bool _isWhiteBackground;
 
   Style get _selectionStyle => widget.controller.getSelectionStyle();
 
@@ -51,7 +51,7 @@ class _ColorButtonState extends State<ColorButton> {
           widget.controller.getSelectionStyle().attributes);
       _isWhite = _isToggledColor &&
           _selectionStyle.attributes['color']!.value == '#ffffff';
-      _isWhitebackground = _isToggledBackground &&
+      _isWhiteBackground = _isToggledBackground &&
           _selectionStyle.attributes['background']!.value == '#ffffff';
     });
   }
@@ -63,7 +63,7 @@ class _ColorButtonState extends State<ColorButton> {
     _isToggledBackground = _getIsToggledBackground(_selectionStyle.attributes);
     _isWhite = _isToggledColor &&
         _selectionStyle.attributes['color']!.value == '#ffffff';
-    _isWhitebackground = _isToggledBackground &&
+    _isWhiteBackground = _isToggledBackground &&
         _selectionStyle.attributes['background']!.value == '#ffffff';
     widget.controller.addListener(_didChangeEditingValue);
   }
@@ -87,7 +87,7 @@ class _ColorButtonState extends State<ColorButton> {
           _getIsToggledBackground(_selectionStyle.attributes);
       _isWhite = _isToggledColor &&
           _selectionStyle.attributes['color']!.value == '#ffffff';
-      _isWhitebackground = _isToggledBackground &&
+      _isWhiteBackground = _isToggledBackground &&
           _selectionStyle.attributes['background']!.value == '#ffffff';
     }
   }
@@ -106,7 +106,7 @@ class _ColorButtonState extends State<ColorButton> {
         : (widget.iconTheme?.iconUnselectedColor ?? theme.iconTheme.color);
 
     final iconColorBackground =
-        _isToggledBackground && widget.background && !_isWhitebackground
+        _isToggledBackground && widget.background && !_isWhiteBackground
             ? stringToColor(_selectionStyle.attributes['background']!.value)
             : (widget.iconTheme?.iconUnselectedColor ?? theme.iconTheme.color);
 
@@ -114,7 +114,7 @@ class _ColorButtonState extends State<ColorButton> {
         ? stringToColor('#ffffff')
         : (widget.iconTheme?.iconUnselectedFillColor ?? theme.canvasColor);
     final fillColorBackground =
-        _isToggledBackground && widget.background && _isWhitebackground
+        _isToggledBackground && widget.background && _isWhiteBackground
             ? stringToColor('#ffffff')
             : (widget.iconTheme?.iconUnselectedFillColor ?? theme.canvasColor);